# Docs/swagger/Geographic area service/area_documentation.yml
openapi: 3.0.3
info:
  title: Geographic area microservice - OpenAPI 3.0
  description: |-
    service to consult and search geographical areas (neighborhoods)
  version: 1.0.0
servers:
  - url: https://calitel.area/api/v1
tags:
  - name: Areas
    description: Everything about geographic areas
paths:
  /area:
    get:
      tags:
        - Areas
      summary: Get Cali neighborhoods
      description: Get Cali neighborhoods (geographical areas)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Area'
        '500':
          description: Failed to get areas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /area/{center_id}:
    get:
      tags:
        - Areas
      summary: Get neighborhoods in Cali by a central 
      description: Obtain neighborhoods in Cali that are served by a central
      parameters:
        - $ref: '#/components/parameters/center_id'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Area'
        '500':
          description: Failed to get areas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    Area:
      type: object
      properties:
        area_id:
          type: integer
        name:
          type: string
        code:
          type: integer
        commune:
          type: integer
  parameters:
    center_id:
      in: path
      name: center_id
      required: true
      description: center id
      schema:
        type: integer